+++ cmd/lint_render_summary_view.go
